requirements.txt
Flask
Flask-Cors
google-generativeai
yt-dlp
youtube-transcript-api